requirements.txt
Flask==3.1.1
Werkzeug==3.1.3
line-bot-sdk==3.17.1
google-generativeai==0.8.0
google-auth==2.40.3
google-cloud-storage
Pillow==11.3.0
PyMySQL==1.1.1
requests==2.32.4
urllib3==2.3.0
python-dotenv==1.1.1
schedule==1.2.2
pydantic==2.11.7
Jinja2==3.1.6
click==8.2.1
pytz==2024.2
cryptography==43.0.3
gunicorn==21.2.0
google-cloud-speech==2.21.0
pydub==0.25.1